Added functionality to display an index of favourited collections.
Pulled the body of the collection index page out of the blade and moved it into a partial so it can be reused for other similar pages (favourites/etc).

Modified how rating restrictions are used on the index to be a bit cleaner using whereNotIn instead of a foreach loop.